[ENH]
- Implemented new grid container class
- compute model can compute gravity
- compute at only modifies grid.custom_grid
- InterpolatorGravity now set tz automatically
- set_grid_object substituted by individual grid types
- set_gravity_interpolation added
- Solution adapted to new grids

[BUG]

theano.tz now gets dtype right
surfaces now are loaded right (vertices and edges are not saved)

+++ gempy/core/theano/theano_graph_pro.py
@@ -209,7 +209,7 @@ class TheanoGraphPro(object):
         self.offset = theano.shared(10.)
 
         # Gravity
-        self.tz = theano.shared(np.empty((0), dtype='float32'), 'tz component')
+        self.tz = theano.shared(np.empty((0), dtype=dtype), 'tz component')
         self.density_matrix = T.vector('density vector')
 
         # Results matrix
